Remove spurrious file
