+++ include/eve/module/real/core/function/regular/generic/add.hpp
@@ -8,8 +8,8 @@
 #pragma once
 
 #include <eve/concept/compatible.hpp>
-#include <eve/detail/function/simd/common/conditional.hpp>
 #include <eve/concept/value.hpp>
+#include <eve/detail/function/conditional.hpp>
 #include <eve/traits/common_compatible.hpp>
 
 namespace eve::detail
