+++ include/eve/module/real/core/function/regular/generic/bit_not.hpp
@@ -9,7 +9,7 @@
 
 #include <eve/concept/value.hpp>
 #include <eve/detail/implementation.hpp>
-#include <eve/detail/function/simd/common/conditional.hpp>
+#include <eve/detail/function/conditional.hpp>
 #include <eve/function/bit_cast.hpp>
 #include <eve/forward.hpp>
